requirements.txt
fastapi
uvicorn
pandas
psycopg2-binary
sentence-transformers
llama-cpp-python
python-dotenv
faiss-cpu